Only build datasets on required ranks
Currently, during the creation of the DataLoader ALL ranks are accessing (open + read) 3 tiny files: the tokenizer file, the blend file (--per-split-data-args-path) & the --per-dataset-sequences-path file.
With this commit, when leveraging the --dataloader-fast-cache-load flag, we only build the datasets on the required ranks. For the ultra run, this will reduce the total amount of I/O by 8, since we are using TP=8 and only building the datasets on tp_rank=0.